# swagger/api-docs.json
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
...
paths:
  /tour/{id}:
    get:
      tags:
        - Tour
      summary: Obtém detalhes de um passeio por ID
      description: Retorna informações detalhadas de um passeio com base no ID fornecido.
      operationId: getById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tour'
          
    put:
      tags:
        - Tour
      summary: Atualiza um passeio por ID
      description: Atualiza as informações de um passeio existente com base no ID.
      operationId: update
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TourDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object


    delete:
      tags:
        - Tour
      summary: Deleta um passeio por ID
      description: Apaga um passeio existente com base no ID.
      operationId: delete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /tour:
    get:
      tags:
        - Tour
      summary: Lista todos os passeios
      description: Retorna uma lista de todos os passeios disponíveis.
      operationId: getAll
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tour'
            
  
    post:
      tags:
        - Tour
      summary: Cria um novo passeio
      description: Cria um novo passeio com base nos dados fornecidos.
      operationId: createTour
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TourDTO'
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
components:
  schemas:
    TourDTO:
      required:
        - agenciaId
        - destino
        - itinerario
        - preco
      type: object
      properties:
        preco:
          type: number
          format: double
        itinerario:
          type: string
        destino:
          type: string
        agenciaId:
          type: string
          format: uuid
...
    Person:
      type: object
      properties:
        id:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        nome:
          type: string
        email:
          type: string
        nascimento:
          type: string
          format: date-time
        tipo:
          type: string
          enum:
            - CLIENTE
            - AGENCIA_VIAGEM
        cpf:
          type: string
    Tour:
      type: object
      properties:
        id:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        agencia:
          $ref: '#/components/schemas/Person'
        preco:
          type: number
          format: double
        itinerario:
          type: string
        destino:
          type: string
...